access validation

+++ src/core/kernel/commands/filesystem/rmdirCommand.ts
@@ -1,17 +1,23 @@
-import { Path, Permissions } from "@thijmen-os/common";
+import { Access, Path, Permissions } from "@thijmen-os/common";
 import { ICommand } from "@ostypes/CommandTypes";
 import { RemoveDirectory } from "@providers/filesystemEndpoints/filesystem";
+import CommandAccessValidation from "@core/kernel/accessValidation";
 
-class rmdirCommand implements ICommand {
+class rmdirCommand extends CommandAccessValidation implements ICommand {
   private props: Path;
 
   readonly requiredPermission = Permissions.fileSystem;
 
-  constructor(props: { path: Path }) {
-    this.props = props.path;
+  constructor(props: Path) {
+    super();
+
+    this.props = props;
   }
 
   public Handle(): void {
+    const validated = this.validateAccess(this.props, Access.w);
+    if (!validated) return;
+
     RemoveDirectory(this.props);
   }
 }
